requirements.txt
streamlit==0.63.0
opencv_contrib_python==4.3.0.36
numpy==1.19.0
Pillow==7.2.0
